moved filter default data from /constants to /helpers/resultsTableConfig
